add usecase
